backend/requirements.txt
# Core Web Framework
fastapi==0.115.8
uvicorn==0.22.0

# Database
sqlalchemy==2.0.12
alembic==1.10.3
psycopg2-binary==2.9.6

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
pydantic-settings==2.0.1
pydantic[email]

# Payment Gateway
paypalrestsdk

# External API & Requests
requests==2.28.2

# Environment & Configuration
python-dotenv==1.0.0

# Validation
#pydantic==2.1.1

# Distance Calculations
geopy==2.3.0

# Async Support
asyncio==3.4.3

# # Testing
# pytest==7.3.1
# httpx==0.24.0

# Logging
loguru==0.7.0

# CORS Handling
python-multipart==0.0.6

pytz

# Background Tasks
celery
redis
numpy
flower
scikit-learn